+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## Next
 
 - recording: capture touch interaction off of main thread to avoid ANRs ([#165](https://github.com/PostHog/posthog-android/pull/165))
+- chore: use getNetworkCapabilities instead of getNetworkInfo to avoid ANRs ([#164](https://github.com/PostHog/posthog-android/pull/164))
 
 ## 3.5.0 - 2024-08-08